Update path total risk function_kelly.cpp

+++ path total risk function_kelly.cpp
@@ -1,7 +1,7 @@
-double totalLineRisk (int Xstart, int Xend, int Ystart, int Yend, int *X, int *Y, int *R, int *P)
+float totalLineRisk (int Xstart, int Xend, int Ystart, int Yend, int *X, int *Y, int *R, int *P)
 {
-	double totalRisk = 0;
-	double length = sqrt(pow(Xstart - Xend,2) + pow(Ystart - Yend,2));
+	float totalRisk = 0;
+	float length = sqrt(pow(Xstart - Xend,2) + pow(Ystart - Yend,2));
 		
 	int times = static_cast<int>(length);   //轉折點個數 
 		
@@ -10,12 +10,12 @@ double totalLineRisk (int Xstart, int Xend, int Ystart, int Yend, int *X, int *Y
 	 
 	for(int t = 1; t <= times; t++)//dot num
 	{
-		double x = Xstart + (Xend-Xstart) / length * t;  //下一個的x座標 
-		double y = Ystart + (Yend-Ystart) / length * t;  //下一個的y座標 
+		float x = Xstart + (Xend-Xstart) / length * t;  //下一個的x座標 
+		float y = Ystart + (Yend-Ystart) / length * t;  //下一個的y座標 
 			
 		for(int b = 0; b < m; b++)//bomb num
 		{
-			double d = sqrt(pow((x - X[b]), 2) + pow((y - Y[b]), 2));   //分割點和炸彈的距離 
+			float d = sqrt(pow((x - X[b]), 2) + pow((y - Y[b]), 2));   //分割點和炸彈的距離 
 				
 			if(R[b] - d > 0)
 			{
